requirements.txt
asgiref==3.8.1
attrs==24.1.0
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
